Initial Events API support.
Decided to add support for both V1 and V2 events, and through a common
interface even though they'll come through a single channel for the time
being.

For now I'm just treating enums and timestamps as strings, but might do
a separate pass to lock them down. We also aren't doing any validation
at the moment, but could be something to consider in the future.

Will add tests (including mocking) in a subsequent PR.