Moving retry handling from the queue processor to `eventsapi`.
This is closer to what we probably should have had from the start:
Invisibly handling retries as middleware (a `RoundTripper`) rather than
requiring that processors handle it.

This also meant moving concepts like `retryable` and corresponding
`success` to the `RetryTransport` rather than making it part of
`eventsapi` response handling.

A next step could be to further pull this out into its own service or
package, but for now it's at least fairly self-contained within
`eventsapi`.